Allo Friendly Fire Values to be float

+++ valo_api/responses/match_history.py
@@ -26,8 +26,8 @@ class MatchPlayerSessionPlaytimeV3(DictStruct):
 
 
 class MatchFriendlyFireV3(DictStruct):
-    incoming: Optional[int] = None
-    outgoing: Optional[int] = None
+    incoming: Optional[float] = None
+    outgoing: Optional[float] = None
 
 
 class MatchPlayerBehaviorV3(DictStruct):
